+++ motivoo-api/src/main/java/sopt/org/motivoo/api/controller/auth/dto/response/OAuthPlatformMemberResponse.java
@@ -1,4 +1,4 @@
-package sopt.org.motivoo.api.controller.user.apple;
+package sopt.org.motivoo.api.controller.auth.dto.response;
 
 
 import lombok.AccessLevel;
