+++ motivoo-external/src/main/java/sopt/org/motivoo/external/client/auth/FeignClientConfig.java
@@ -1,8 +1,9 @@
-package sopt.org.motivoo.external;
+package sopt.org.motivoo.external.client.auth;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @EnableFeignClients(basePackageClasses = FeignClientConfig.class)
 public class FeignClientConfig {
+
 }
