+++ motivoo-external/src/main/java/sopt/org/motivoo/external/client/auth/apple/AppleClient.java
@@ -1,8 +1,7 @@
-package sopt.org.motivoo.external;
+package sopt.org.motivoo.external.client.auth.apple;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
-import sopt.org.motivoo.external.auth.apple.response.ApplePublicKeys;
 
 @FeignClient(name = "apple-public-key-client", url = "https://appleid.apple.com/auth")
 public interface AppleClient {
