+++ motivoo-external/src/main/java/sopt/org/motivoo/external/client/auth/apple/ApplePublicKeys.java
@@ -1,6 +1,6 @@
-package sopt.org.motivoo.external.auth.apple.response;
+package sopt.org.motivoo.external.client.auth.apple;
 
-
+import static sopt.org.motivoo.common.advice.CommonExceptionType.*;
 
 import java.util.List;
 
@@ -9,7 +9,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import sopt.org.motivoo.common.advice.BusinessException;
-import sopt.org.motivoo.domain.user.exception.UserExceptionType;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
@@ -23,7 +22,7 @@ public class ApplePublicKeys {
                 .stream()
                 .filter(k -> k.getAlg().equals(alg) && k.getKid().equals(kid))
                 .findFirst()
-                .orElseThrow(() -> new BusinessException(UserExceptionType.INVALID_APPLE_PUBLIC_KEY));
+                .orElseThrow(() -> new BusinessException(INVALID_APPLE_PUBLIC_KEY));
     }
 
 }
