+++ motivoo-external/src/main/java/sopt/org/motivoo/external/client/auth/apple/service/AppleClaimsValidator.java
@@ -1,4 +1,4 @@
-package sopt.org.motivoo.domain.auth.service.apple;
+package sopt.org.motivoo.external.client.auth.apple.service;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
