+++ motivoo-external/src/main/java/sopt/org/motivoo/external/client/auth/apple/service/EncryptUtils.java
@@ -1,7 +1,7 @@
-package sopt.org.motivoo.domain.auth.service.apple;
+package sopt.org.motivoo.external.client.auth.apple.service;
 
 
-import static sopt.org.motivoo.domain.user.exception.UserExceptionType.*;
+import static sopt.org.motivoo.common.advice.CommonExceptionType.*;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
